module_2(fix): rates converter refactoring
