+++ src/components/home-page-blocks/news/news-carousel/news-carousel.module.scss
@@ -56,7 +56,7 @@
   &:disabled {
     background-color: transparent;
     border: 1px solid colors.$blue;
-  
+
     svg path {
       stroke: colors.$dark;
     }
